% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gstar.R
\name{gstar}
\alias{gstar}
\title{Fit Generalized Space-Time Autoregressive Model}
\usage{
gstar(x, weight, p = 1, d = 0, est = "OLS")
}
\arguments{
\item{x}{a dataframe, matrix or xts/ts object that contain time series data.}

\item{weight}{a spatial weight ncol(x) * ncol(x) with diagonal = 0.}

\item{p}{an autoregressive order, value must be greater than 0.}

\item{d}{a lag differencing order, value must be greater than 0.}

\item{est}{estimation method, currently only OLS available, another estimation will be added later.}
}
\value{
gstar returns output like lm, use function \code{\link{summary}}to obtain the summary coefficients and others. The detail are shown in the following list :
\itemize{
\item coefficients - a named vector of coefficients.
\item AIC - A version of Akaike's An Information Criterion (the calculation similiar with aic in \emph{lm} method )}
}
\description{
gstar function return the paramter estimatation of Generalized Space-Time Autoregressive Model.
}
\examples{
library(gstar)
library(xts)
data("LocationCPI")

x = xts(LocationCPI[, -1], order.by = as.Date(LocationCPI[, 1]))
weight = matrix(c(0, 1, 1, 1,
                1, 0, 1, 1,
                1, 1, 0, 1,
                1, 1, 1, 0), ncol = 4, nrow = 4)

#the sum of weight is equal to 1 every row.
weight = weight/(ncol(x) - 1)

fit <-  gstar(x, weight = weight, p = 1, d = 0, est = "OLS")
summary(fit)

performance(fit)

predict(fit)

plot(fit)
}
\references{
Budi Nurani Ruchjana, Svetlana A. Borovkova and H. P. Lopuhaa (2012), \emph{Least Squares Estimation of Generalized Space Time AutoRegressive (GSTAR) Model and Its Properties}, The 5th International Conference on Research and Education in Mathematics AIP Conf. Proc. 1450, 61-64.
}
\seealso{
\code{\link{summary}} for summarises the model that has been built. Also use \code{\link{predict}} to predict model to testing or new data.
}
