% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Loc}
\alias{Loc}
\title{Coordinate of several region In Indonesia}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{City}{Name of region/city}
  \item{lattitude}{The lattitude coordinate of each location}
  \item{longitude}{The longitude coordinate of each location}
}}
\usage{
data(Loc)
}
\description{
A dataset containing the coordinate several region In Indonesia i.e Semarang, Surakarta, Tegal and Purwokerto.
}
\keyword{datasets}
