% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LocationCPI}
\alias{LocationCPI}
\title{Consumer Price Index (CPI) in several region In Indonesia}
\format{A timeseries data frame with 105 rows and  5 variables:
\describe{
  \item{Date}{date of CPI, monthly}
  \item{Purwokerto}{The CPI of Purwokerto region}
  \item{Surakarta}{The CPI of Purwokerto region}
  \item{Semarang}{The CPI of Purwokerto region}
  \item{Tegal}{The CPI of Purwokerto region}
}}
\source{
\url{https://www.bps.go.id/}
}
\usage{
data(LocationCPI)
}
\description{
A dataset containing the Consumer Price Index (CPI) in several region In Indonesia i.e Semarang, Surakarta, Tegal and Purwokerto, it is timeseries data with monthly periodicity fromJan 2016 untill Sep 2014
}
\keyword{datasets}
