% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfomance.R
\name{performance}
\alias{performance}
\title{Calculate performance of prediction or forecasting}
\usage{
perfomance(object, testing = NULL)
}
\arguments{
\item{testing}{a dataframe or matrix or xts object that contain testing data (outsample data). \strong{Please be noted, if you use differencing in the model estimation, you do not need difference your data because we already cover that in this function}}

\item{object}{an object of class "gstar".}
}
\value{
\itemize{
\item MSE fol all data - Mean Square Error for all the data combined
\item MSE fol each location - Mean Square Error for each spatial location
\item MAPE fol all data - Mean Absolute Percentage Error for all the data combined
\item MAPE fol each location - Mean Absolute Percentage Error for each spatial location }
}
\description{
Calculate performance of prediction or forecasting
}
