% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gstar}
\alias{predict.gstar}
\title{Predicting the gstar model}
\usage{
predict(object, n)
}
\arguments{
\item{n}{The number of steps ahead for which prediction is required.}

\item{object}{an object of class "gstar".}
}
\description{
Predicted values based on gstar model object
}
