% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gstar.R
\name{summary.gstar}
\alias{summary.gstar}
\title{Summarizing Generalized Space-Time Autoregressive Fits}
\usage{
\method{summary}{gstar}(model, digits = max(3L, getOption("digits") -
  3L), ...)
}
\description{
This function are simmilar with summary to "lm" or "glm" object.
\itemize{
\item coefficients - a named vector of coefficients.
\item AIC - A version of Akaike's An Information Criterion.
}
}
