% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gstar.R
\name{plot.gstar}
\alias{plot.gstar}
\title{Plotting the gstar object}
\usage{
\method{plot}{gstar}(x, n_predict = NULL, testing = NULL, ...)
}
\arguments{
\item{x}{an object of class "gstar".}

\item{n_predict}{The number of steps ahead for which prediction is required.}

\item{testing}{The outsample or testing data to be plotted.}

\item{...}{Other arguments}
}
\description{
plotting the gstar object
}
