% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.gstar}
\alias{predict.gstar}
\title{Predicting the gstar object}
\usage{
\method{predict}{gstar}(object, n = NULL, ...)
}
\arguments{
\item{object}{an object of class "gstar".}

\item{n}{The number of steps ahead for which prediction is required.}

\item{...}{another arguments}
}
\description{
Predicted values based on gstar object object
}
