% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_gstar.R
\name{summary.gstar}
\alias{summary.gstar}
\title{Summarizing Generalized Space-Time Autoregressive Fits}
\usage{
\method{summary}{gstar}(object, ...)
}
\arguments{
\item{object}{an object of class "gstar".}

\item{...}{further arguments passed to or from other methods.
\itemize{
\item coefficients - a named vector of coefficients.
\item AIC - A version of Akaike's An Information Criterion.
}}
}
\description{
This function are simmilar to summary of "lm" or "glm" object.
}
